Rename fileutils package naar util. Gitignore toegevoegd. Readme toegevoegd.
