+++ README.md
@@ -1,0 +1,7 @@
+#Fotool applicatie
+
+Applicatie om bewerkingen op fotobestanden uit te voeren. In eerste instantie bedoeld om fotobestanden te kunnen
+hernoemen op basis van de EXIF-data in de foto's. Dit is bijvoorbeeld handig om de foto's van meerdere camera's
+een nieuwe naam te geven met een chronologisch oplopend volgnummer.
+
+Dit is een command-line applicatie zonder GUI.
